Moved virtualizd autocomplete to a component
